bump minor version number (#2)
* add soft breaks to New Line logic

In our application we sometimes get substation alpha authors that put n at
the end of each line in the dialogue. Usually we would expect that they would
use N instead.  So we will add that to the conditions checked by the parse-
newLine function so that this doesn'
t cause a breaking behavior of actually printing out the literal n at
the end of each line of dialogue.

* bump version minor number